implemented Snake with arrow controls and drawing

+++ src/view/tile.rs
@@ -9,14 +9,14 @@ pub enum TileType {
 
 #[derive(Copy, Clone)]
 pub struct Tile {
-    pub x: u8,
-    pub y: u8,
+    pub x: i32,
+    pub y: i32,
     pub ttype: TileType
 }
 
 
 impl Tile {
-    pub fn new(posx: u8, posy: u8, tile_type: TileType) -> Self {
+    pub fn new(posx: i32, posy: i32, tile_type: TileType) -> Self {
         Self {
             x: posx,
             y: posy,
